refactor(web): 拆 Agent 仪表盘为两个独立页面
Agent 仪表盘把"两个 Agent 概览 / 主 Agent 上下文 / 调度任务"三个东西塞在
一个页面里，但实际有用的只有后两个。概览页堆了一堆 state tree 时代的字
段（focusedStateId / stateStack / 可用 invoke 工具等），等 App 体系扩
开后这些字段会陆续被替换，留着只是误导。

拆分：

- 新页面 /main-agent-context：主 Agent 最近上下文
  - 复用 /agent-dashboard/current 接口，只渲染 root agent 的 recentItems
  - 后端 schema 暂不动，等后续整体清理底层 dashboard query 时再瘦身
- 新页面 /scheduler-tasks：调度任务
  - 完整搬走原"调度任务"tab 的 UI 与逻辑
  - useSchedulerTasks 从 agent-dashboard/ 移到 scheduler-tasks/

删除：

- apps/web/src/pages/agent-dashboard/ 整个目录（含 AgentDashboardPage 与
  useAgentDashboardSnapshot）
- 默认路由从 /agent-dashboard 改为 /main-agent-context
- 导航菜单 "Agent 仪表盘" 改为两条 "主 Agent 上下文" + "调度任务"

为什么概览 tab 直接删而不是搬迁：

- "两个 Agent 概览" 大部分内容是 state tree 时代的存量字段（focusedStateId、
  stateStack、availableInvokeTools、子节点列表）。App 体系下这些字段要么
  概念已变要么会消失，再展示反而误导
- providers / loop state / 最近 LLM 调用 / 最近错误 这些有信息价值的字
  段，分散在 LLM 历史 / 应用日志 / metric 图表三个页面里都能查，没必要
  在概览里重复列
- 后端 /agent-dashboard/current 接口暂时保留（context 页面用），等清理
  底层 schema 时一并瘦身

439 / 439 测试通过；frontend build / typecheck / lint / format 全绿。
